# Conflicts:
#   backend/sswm/src/main/java/com/ground/sswm/userStudyroom/controller/UserStudyroomController.java
#   backend/sswm/src/main/java/com/ground/sswm/userStudyroom/domain/UserStudyroomRepository.java
#   backend/sswm/src/main/java/com/ground/sswm/userStudyroom/dto/OnAirResDto.java
#   backend/sswm/src/main/java/com/ground/sswm/userStudyroom/dto/UserStudyTimeResDto.java
#   backend/sswm/src/main/java/com/ground/sswm/userStudyroom/dto/UserStudyroomReqDto.java
#   backend/sswm/src/main/java/com/ground/sswm/userStudyroom/service/UserStudyroomService.java
#   backend/sswm/src/main/java/com/ground/sswm/userStudyroom/service/UserStudyroomServiceImpl.java

+++ backend/sswm/src/main/java/com/ground/sswm/auth/domain/AuthRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface AuthRepository extends JpaRepository<Auth,Integer> {
-    Auth findByUserId(int userId);
+    Auth findByUserId(Long userId);
 }
